+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
-    "git.ignoreLimitWarning": true
+    "git.ignoreLimitWarning": true,
+    "python.pythonPath": "${workspaceFolder}\\venv\\Scripts\\python.exe"
 }
