+++ css/header.css
@@ -15,8 +15,8 @@ header>form {
     height: 40px;
 }
 
-header>form>input{
-border-radius:5px;
+header>form>input {
+    border-radius: 5px;
 }
 
 header>.top-spacer {
@@ -24,16 +24,40 @@ header>.top-spacer {
     flex: 1;
 }
 
+header>.btn1 {
+    margin-top: 12px;
+    background-color: blue;
+    margin-left: 130px;
+    width: 91px;
+    height: 40px;
+    border: none;
+    border-radius: 5px;
+    color: #fff;
+}
+
+header>.btn2 {
+    margin-right: 30px;
+    margin-top: 12px;
+    background-color: green;
+    margin-left: 3px;
+    width: 91px;
+    height: 40px;
+    border: none;
+    border-radius: 5px;
+    color: #fff;
+}
 
 header>.dropdown {
     display: flex;
-    flex:1;
+    flex: 1;
     justify-content: flex-end;
-    margin:10px 20px 0 0;
-    background-color:rgb(255, 255, 255);
-    min-width:50px;
-    max-height:50px;
-    border-radius:50px;
+    margin: 10px 20px 0 0;
+    background-image:url("/img/top-avatar.png");
+    background-repeat: no-repeat;
+    min-width: 50px;
+    max-height: 50px;
+    border-radius: 50px;
+    cursor: pointer;
 }
 
 header>.dropdown>.dropdown-content {
@@ -41,7 +65,7 @@ header>.dropdown>.dropdown-content {
     position: absolute;
     background-color: #f9f9f9;
     min-width: 100px;
-    top:65px;
+    top: 65px;
     box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
     padding: 12px 16px;
 }
@@ -49,6 +73,7 @@ header>.dropdown>.dropdown-content {
 header>.dropdown:hover>.dropdown-content {
     display: block;
 }
+
 
 .dropdown-content a {
     background-color: #eee;
@@ -67,7 +92,8 @@ header>.dropdown:hover>.dropdown-content {
     flex-direction: column;
     flex: 1;
     max-width: 20%;
-    background-color: #eee;z
+    background-color: #eee;
+    z
 }
 
 .vertical-menu>h2 {
@@ -104,30 +130,13 @@ header>.dropdown:hover>.dropdown-content {
     color: white;
 }
 
-header>.btn1{
-    margin-top:12px;
-    background-color:blue;
-    margin-left:130px;
-    width:91px;
-    height:40px;
+.form>form>div {
+    margin-bottom: 10px;
 }
 
-header>.btn2{
-    margin-right:30px;
-    margin-top:12px;
-    background-color:green;
-    margin-left:3px;
-    width:91px;
-    height:40px;
-}
-
-.form>form>div{
-    margin-bottom:10px;
-}
-
-.form>form>div>button{
-    background-color:blue;
-    margin-left:130px;
-    width:91px;
-    height:40px;
+.form>form>div>button {
+    background-color: blue;
+    margin-left: 130px;
+    width: 91px;
+    height: 40px;
 }
