+++ dashboard.html
@@ -17,7 +17,6 @@
         <div style="flex:20;"></div>
         <div class="dropdown">
             <div class="dropdown-content">
-                <a href="#" class="active">My account</a>
                 <a href="login.html">signout</a>
             </div>
         </div>
