+++ question_answered.html
@@ -35,7 +35,7 @@
             </form>
             <a href="dashboard.html">General posts</a>
             <a href="question.html" class="active">My posts</a>
-            <a href="profile.html">Profile</a>
+            <a href="update-profile.html">Profile</a>
         </aside>
         <section class="content">
             <div class="posts">
