+++ update-profile.html
@@ -26,7 +26,7 @@
 
 
 
-    <div class="container ">
+    <div class="container">
         <!--main content-->
         <aside class="vertical-menu">
             <h2>menu</h2>
@@ -38,60 +38,48 @@
             <a href="profile.html" class="active">Profile</a>
         </aside>
 
-        <section class="content container-pro">
-            
+        <section class="content">
 
-            <form id="profile-form" action="#">
-
-                <div class="propic">
-<h2>Edit Profile</h2>
-                    <div class="column">
-                        <img src="img/man-avatar.png" alt="Snow" style="width:100%">
-                    </div>
-                    <input type="file" id="upload">
-                        <br>
-<h3>About you</h3>
-                    <textarea></textarea>
-                </div>
+            <form action="#">
                 <div class="contain">
-
+                    <h1>Edit your profile</h1>
+                    <hr>
+<p>
+    Public information</p>
                     <div>
-                        <span class="label">
-                            Display Name
-                        </span>
-                        <input type="text" placeholder="e.g  John Doe" name="email" required>
+<i class="label">
+                            <b>Display Name</b>
+                        </i>
+                        <input type="text" placeholder="e.g John Doe" name="email" required>
                     </div>
 
                     <div>
-                        <span class="label">
-                            Email
-                        </span>
-                        <input type="email" placeholder="e.g  name@example.com" name="email" required>
-                    </div>
-
-                    <div>
-                        <span class="label">
-                            Phone
-                        </span>
-                        <input type="text" placeholder="e.g  070 - xxx - xxx" name="email" required>
+<i class="label">
+                            <b>Email:</b>
+                        </i>
+                        <input type="email" placeholder="e.g name@example.com" name="email" required>
                     </div>
                     <div>
-                        <span class="label">
-                            Current password
-                        </span>
-                        <input type="password" placeholder="********" name="email" required>
+<i class="label">
+                            <b>Name</b>
+                        </i>
+                        <input type="text" placeholder="Enter Email" name="email" required>
                     </div>
+<div>
+<i class="label">
+        <b>Current password</b>
+    </i>
+    <input type="password" placeholder="********" name="email" required>
+</div>
                     <div>
-                        <span class="label">
-                            Repeat Password
-                        </span>
-                        <input type="password" placeholder="********" name="psw-repeat" required>
+                        <i class="label">
+                            <b>Repeat Password</b>
+                        </i>
+                        <input type="password" placeholder="Repeat Password" name="psw-repeat" required>
                         <hr>
                     </div>
-
                     <button type="submit" class="registerbtn">Save</button>
                 </div>
-
             </form>
 
         </section>
